03_End-to-End_RAG/requirements.txt
# Session 03: End-to-End RAG System Dependencies
# Core FastAPI and server dependencies
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6

# RAG and AI dependencies
openai>=1.0.0
python-dotenv>=1.0.0

# PDF processing
PyPDF2>=3.0.0

# Data processing and utilities
pydantic>=2.0.0
typing-extensions>=4.0.0

# aimakerspace library (as required by homework)
# Note: This should be installed from the local aimakerspace folder
# pip install -e ../aimakerspace
